Clean up ScaleEditor.tsx. Remove ScaleEditorInput.tsx, start fresh with ScaleNoteInput.tsx and create new classes for base note and octave note. Narrowed scope for props in ScaleEditor (more to come).
Behavior changes:
A user can convert from a cents to a ratio only if they just converted their ratio to a cent value and have not since edited it.
If a user enters '100' and the previous value was a cent value, a '.0' will be appended to keep the cent value status. Before we turned it to a ratio value.